Bugfix: fix searching and not searching dates, when delta should be 0 and when delta can be 0 or other
When I implemented the search, I had to disable the date delta to allow for all dates to be searched.
This broke the access control checks, which require the delta = 0.

As a solution to both, specifically add the delta = 0 check for all non-search list loading.
The downside to having no restrictions on delta is that the row count from the db_select will possibly not be accurate.
This inaccuracy happens because multiple rows for a single event will be loaded.
I am unaware of a way to do this properly via the SQL language.
(It would require relative distincts with conditional column merges.)