Update: Change the caching of the banner data into storing it into tables for better performance and to improve database design
In particular, this fixes an oversight where buildings are allowed to be in multiple locations.
This is likely why they called 'locations' 'classifications' in the banner database.

This new approach is more efficient in that instead of using the drupal serialized caching, the data can now simply be queried and joined as needed.

To ensure consistency with the banner database, it will be constantly synchronized via cron runs.
Furthermore, to help prevent data loss, 'disabled' columns were added for anything that no longer appears in the banner database.

This allows the information on this system to remain, for archiving purposes, should the banner information be no longer available.