Bugfix: the search listing was not providing all of the results
It appears that due to the multi-date values (and times), too many results were being generated.
To prevent this from happening, do not select the date values and assign the query as distinct.

This will allow for all of the values to be searched, but duplicate rows will not be presented.