Cleanup: usage fee wave should be shown on the WAVE decision
I may have missed something when I moved the DECISION_WAVE into the DECISION_USAGE_FEE as requested.
In certain cases, it seems that the log displays only the WAVE and not the USAGE FEE (I don't remember if that is intentional).
Be sure to handle the case where the fee wave is display in the log whether it is either a DECISION_USAGE_FEE or a DECISION_WAVE.