Bugfix: fix more permissions issues
I was sloppy when I rushed in permission changes before the meeting.
The standard view access should not depend on whether or not the current user is allowed to review.
Reviewers can view the event whether or not it can be reviewed.

I also noticed some mistakes in the cases where $event_id is null.
When $event_id is null, then the access check is whether or not they can
'view' or 'review' and not whether or not they can 'view' or 'review' a specific event.