Update: revert behavior of WAVED to allow for reviewers to specify fees regardless
The original behavior had made the WAVED status not restrect the usage fees specified by others.
This was specifically requested to not happen and so the WAVED process was completely redesigned to lock out usage fees when waved.

They have now decided that they want reviewers to specify usage fees regardless of waved status.
This essentially reverts to the previous behavior.

However, I did not want to completely revert in case they changed their minds.
I removed only the relevant code, this commit can be undone if/when they changed their minds.