Update: switch to new database users and names (partial)
This is a partial switch because after I renamed the main database to mfcs, I realized that it should instead be fcs.

The testing system will be renamed to fcs and once the testing system is up and okay, then I will transition the main system to fcs for consistency purposes.