Bugfix: workflow could get stuck when reviewers are modified
When reviewers are modified such that the last reviewers left ro review a request are the reviewers removed, the request can never move onto the next step.
Solve this by providing manager, who change the reviwers, to recheck the requests.
If the recheck identifies the problem, it will immediately push the request to the next step.