Update: disable organizer for the time being
The ical specification has made the ORGANIZER field conditional to client-side situations.
There is no way to determine this so whether or not the ical file is compliant to the standards is not determined by the server but instead by the client.

Which means that from one user to the next, the same exact ical file could be compliant for one user and non-compliant for another.
IMHO, such a conditional wording should never be specified in an exchange format standard.