Update: change day request positioning algorithm
Instead of doing an simple alternating, determine whether or not the requests need to be condensed or not.
Set a min/max.

Each request is assumed to be about 156px.
The assumed width is 800px.
The allowed number of requests per row before condesning begins is: floor(800/156) = 5.

Once there are more than 5, determine the number and condense as little as possible.
Thus, at 6 requests, the condensed with is 133 instead of 156 (or 800/6).