Update: limit length of ical lines to 75 chars in accordance to specifications recommendations
The standard states that the character length is 75 characters as a _recommendation_.
I can forsee badly designed clients that interpret this as a _requirement_ instead of a recommendation.
Section 4.1 Content Lines of the standard uses the words "Should Not" and does not use "Must Not".
This is what I interpret as a _recommendation_ and not a _requirement_.

Therefore clobber up the code to make it more likely to be compatible.