Bugfix: access check for view was not checking all cases
When I rewrote the access code I accidentally excluded a case where a user has multiple roles.

Do not exit once the first set of roles are processed, check the others.
If a request id is confirmed to be allowed, then remove it from the ids for the subsequent checks.

This appears to only effect the mfcs_page_request_access_operation_view() function.