Bugfix: drupal user to oracle user code not up to date
It looks like I changed the local database structure for the oracle to drupal tables and I overlooked some necessary changes.
Make sure to change the the banner_id to user_id where necessary (and vice-versa).

It also looks like that with the current database design, new users who did not exist locally during a oracle sync will not have their data available.
This is a problem, so a follow up commit should deal with talking to the oracle database on user creation.

If the oracle database is unavailable, then the user will have to wait till the next sync to have their data made available.