Bugfix: fix use cases of view mode and try to stay on review view mode if possible
When the view mode was changed from an integer to a string (represented by defines), some use cases were not updated.

If the view mode is 'review' which is the operations view mode, check to see if the user still has review access.
If the user still has review access, stay on the review page, otherwise go to the normal view mode.