Bugfix: make sure revision gets logged only once
The table mfcs_request_revisions has a date and a revision number associated with it.
There should be only 1 revision number per instance of time in terms of unix timestamp.

Everything committed during that instance in time is considered part of the same revision.

It looks like I missed a spot where I check to make sure nothing was already committed within this session.

TODO: it would be a good idea to also add a user_id to the revision log so that it could be specific to the instance in time and the user.
(to handle in the hopefully rare case that multiple users make a change at the exact same time in terms of unix timestamp.)