Update: allow searching for created on and updated on dates
If I had planned this out ahead of time, I could have gotten a good implenetation.

It is now possible to search by created and updated date, but it has fewer field options that the other date search fields.
This is because the 'Day' and 'Month' searches aren't possible if there is no value_month or value_day equivalent for the fields.

This might still be possible via SQL date functions (if the standard defines any?) or database-specific date functions.
I have not yet researched these and I do not have the time for this.

The other mer field 'user_id' will not be impleted yet given its own different set of complexities.