Update: use history request status instead to only show delete markup as needed
Determing the status via the $request variable is proper, but the end result is possibly visible markup effects.
Instead of tweaking CSS/markup, only display the deleted status information for the specific commit that the delete happened.

With this change, the variable must be first checked for existence to prevent the problem solved by the previous commit.