Update: improve performance by reducing number of loops associated with dates
Have the load listing fetch associated by the request id.
- This 'appears' to remove a loop, but all it does is change where a loop happens (inside the fetchAllAssoc function).

Use the array keys of the id associated array to load the dates via a single query instead of a loop of queries.
- This increases performance by reducing 1 loop (let the database, which is optimized for looping handle the request ids)..
- This increases performance by making fewer database calls.
- This looses performance by requesting more information (the request_id is now included).
- Overall this is a net gain as the cost of loading the request_id appears to be trivial.

This essentially creates 1 fat query out of X number of smaller queries.
I believe the biggest performance gain is a result of reducing the overhead of building and calling a query.