Bugfix: if all requests have been accessed checked, don't continue processing an empty array
This was discovered on the search page where it is most likely to have already performed an access check on a single request.
- Specifically, this happens when using both simple and advanced search at the same time.

If all requests have been processed, then the request_id array will end up being empty.
An empty array will be passed to numerous functions, causing errors.

The fix is simply to return all of the processed results when there are no remaining request_ids left to process.