Bugfix: date based searching problems
Cleanup the syntax so that the month based variables use the word month and not the word day.

Make the location state a reference so that changes performed during the validate can propogate to the submit function.
- This allows things like 'this month' to work on 'y/m' date searches.

Remove unnecessary and incorrect checks against the 'y/m' value.

The 'y/m/d' and 'y/m' is not properly handling the <, <=, >, and >= operators.
- Because timestamps are dates and time, handling date only search must work with all times for the appropriate dates.