Bugfix: replace REQUEST_TIME with mfcs_instance()
mfcs_instance() allows for having a dynamic/updateable REQUEST_TIME.

This is necessary when there are multiple operations performed within the same php execution instance, but the operations are mutually exclusive.

When using REQUEST_TIME, if the first operation takes 3 days, the second operation would report itself as 3 days in the past instead of when the first request finished.
When using mfcs_instance(), the same situation could be handled either the same day of the first operation or the day the first operation finished its work.