Update: improve performance by rewritting 'current' revision structure
The current design works well, but has a significant performance hit for arbitrarily large numbers of requests.
This is the result of too many joins required to select a given row.
Simplify the design by putting the entire mfcs_current_* collection of tables into a single table.

Doing this reduces a potential 18 joins to 1 join.