Bugfix: mfcs_load_listing() load problems and old code issues
When the fields that were once not part of 'mer' added to the 'mer' table, the search processing code was not completely updated to handle them.
This resulted in loading unneccessary tables and was not properly selecting the appropriate columns.

Some of the old code that remained after changes didn't need to be used as is.
That code has been cleaned up and replaced.