Update: improve accessibility by making sure svg id tags are unique
The svg tags are supposed to be self-contained, but that becomes a grey area when it comes to inline svg.
To ensure that there never is a problem, make sure that the inline svg has unique enough names on the id tags to help reduce the chances that the parent HTML page also has no id tag conflict.

The bars inside the svg were not accessible because the id tags were all the same.