Bugfix: used 'already_joined' names when sorting and remove usage of 'mfcs_view_dates_asc_distinct' view
The already_joined names should be used to prevent selecting rows that may not exist.

The view 'mfcs_view_dates_asc_distinct' had solved significant performance problems in the past, but apparently with the numerous changes to the code it appears to now be causing the performance problem.
I am leaving the view present in case it needs to be re-added due to some unexpected issue.