Update: search module changes and fix handling of search fields with count query
Remove simple search from the load listing function and manually handle simple search.

The count query was failing when search fields were being used because the tables were not being joined.
To fix this, the count query had to be passed to the search building function.
While doing this, I made both the count query and normal query as optional arguments.
- This could cause usage problems as the operand is still being generated but joins are made only to passed query objects.

This seems to fix the weird placeholder issue (from the mcneese module repository):
- 7fb63e7 Workaround: drupals db api overwrites its ouwn placeholders