Bugfix: day calendar is presenting incorrect dates
There is a major screwup in the logic where I did an extra foreach loop.
The deltas for the date should match the deltas for the time start and time stop.
The code, however, had an extra loop through the deltas all over again inside the date delta.
This caused a time entry to exist for each delta such that:
- [ Date 0 ] [ Time 0 ] = correct
- [ Date 0 ] [ Time 1 ] = incorrect
- [ Date 0 ] [ Time 2 ] = incorrect
- [ Date 1 ] [ Time 0 ] = incorrect
- [ Date 1 ] [ Time 1 ] = correct
- [ Date 1 ] [ Time 2 ] = incorrect
- [ Date 2 ] [ Time 0 ] = incorrect
- [ Date 2 ] [ Time 1 ] = incorrect
- [ Date 2 ] [ Time 2 ] = correct

The fix now uses the same delta without an extra loop, producing:
- [ Date 0 ] [ Time 0 ] = correct
- [ Date 1 ] [ Time 1 ] = correct
- [ Date 2 ] [ Time 2 ] = correct

This also means that there is a performance improvement with this fix.