Update: add additional submit fallback to form cases
I had made a patch to drupal core that allows the submit functions to go back to the form for any specified reason.
Utilize that functionality to go back to the form when there is a submit error.
This was already applied to a number of cases before this change.
This change handles the cases that were overlooked and should be returning to the form on submit error.