Security: deny access to users who are not in appropriate Portal/Banner roles
If users are not an 'Administrator', a 'Manager', a 'Reviewer', or a 'Requestor', then no access should be granted.
The system already has access control in place, but what is not being checked is when the user lacks any of the 4 Portal/Banner user roles such as 'Student' or 'Faculty/Staff'.
If the user is not in one of the exception roles such as 'Camps' or 'Externals', also deny the user access.

Provide a message on the dashboard when the user has no access whatsoever.

Note: these are hard-coded values, a more dynamic approach should be looked into should there be time available.
Note: it would still be a good idea to auto-block drupal user acounts for users who are not in the appropriate 4 Portal/Banner roles.