Update: add logging support for cancelling and similar to the review log
The cancel and uncancel operations are not logged to the Review Log.
This adds them to the review log.

While working on this I added the support for adding "Closed" status to the Review Log.
- Currently commented out.

The system user should never have any reviewer_classification.
- Make sure by forcing the reviewer_classification variable to be empty.

The cancel/uncancel support was added by manually calling mfcs_workflow_step_review().
- The e-mails being sent out by the mfcs_workflow_step_review() are ignored because the cancel/uncancel already handles sending out e-mails.
- It would be more consistent to have mfcs_workflow_step_review() handle the e-mailing, but that would require too many changes.

ticket: fcs-11