Bugfix: after amendment, venue coordinator cannot review request
This seems to be an issue due to the caching of request data for performance reasons.
The main table, mfcs_requests, contains the most commonly used columns so that joins can be avoided.
This functionality was added some time after the amendement feature was implemented.

When the performance improvements were made, updated the table was overlooked.
Make sure the appropriate additional information now gets populated.

ticket: fcs-28