Feature: add troubleshooting support
There needs to be a way for managers and administrators to automatically fix all inconsistent content.
Some of this can be auto-detected and auto-fixed and this is what troubleshoot provides.

To fully implement this, I had to alter the history page, some themeing, and add additional GPG functions.
- This utilizes functionality that was made available but never implemented (revision log messages).

It turns out that under certain cases the revision log was not being updated.
- Cases that I noticed will be fixed in a follow up commit.

The following SQL query must be run on an already installed system:
  start transaction;

  create table mfcs_log_troubleshoot (
    id      bigint not null,
    code    bigint not null,
    user_id bigint not null,
    date    bigint not null,

    primary key (id),
    constraint user_id_fkey foreign key (user_id) references users(uid) on delete restrict,

    constraint mfcs_log_troubleshoot_id_check check (id >= 0),
    constraint mfcs_log_troubleshoot_code_check check (code >= 0),
    constraint mfcs_log_troubleshoot_user_id_check check (user_id >= 0),
    constraint mfcs_log_troubleshoot_date_check check (date >= 0)
  );

  create sequence mfcs_log_troubleshoot_id_seq owned by mfcs_log_troubleshoot.id;

  alter table mfcs_log_troubleshoot alter column id set default nextval('mfcs_log_troubleshoot_id_seq'::regclass);

  alter table mfcs_log_troubleshoot owner to fcs_user;

  commit transaction;

ticket: fcs-28