Bugfix: review log not properly showing classifications yet to review
Under certain conditions the classifications that have not yet reviewed are not showing up in the review log screen.

When changing reviewers and reviewer classifications requests that are in the review process can get a bit messy.
It becomes very unclear who is supposed to review what after a change.
Without displaying the classifications, a request might appear to be stuck when it is otherwise not stuck.

There are two distinct functions that are almost identical:
- mfcs_get_reviewers_yet_to_review()
- mfcs_check_if_requests_are_fully_reviewed()

The primary difference is the return value.
Merge those two functions into the following new function:
- mfcs_get_requests_current_review_decisions()
- This new function now accepts NULL as the review revision to allow for auto-loading the current review revision.