Bugfix: ERROR: value too long for type character varying(127) in McNeeseFacilitiesUseMail (mfcs_log_emails table)
I noticed that the title of the request is added to the subject.
The title of the request has a max length of 255.
The e-mail subject has a size of 127.

Bump the e-mail subject size to 384 (because other information gets appended along with the title).

Also bump the destination column to allow for larger e-mail naames (from 127 to 255).

Make sure that the code can never go past these harm limits by wrapping the variables in substr() calls.

The following SQL queries have to be performed along with this commit:
  alter table mfcs_log_emails alter COLUMN destination type character varying(255);
  alter table mfcs_log_emails alter COLUMN title type character varying(384);