Bugfix: non-required use fees did not always nullify when some values were NULL
The use fees requires all or none to be specified.
Because default values are requested to be supplied for amount, this causes a conflict.

To avoid that conflict a check is performed later in the validation process to see if any of the values are null.
If those values are NULL (or strlen == 0), then the entire use fee group should be set to NULL.

The problem appears when one value is non-null before the fields with NULL values and a break is made in the loop.
This prevents the for loop from finding other NULL values to handle accordingly.

Another problem appears when the '#default_value' on the form is either not specified or is also NULL.
When unspecified or NULL, the code for assigning NULL values is unintentionally skipped.
To prevent skipping, when '#default_value' is unspecified or NULL, then only check the appropriate 'form_state' value for NULL.