Cleanup: remove deprecated ajax support on the edit pages
There was an attempt early in the process to use ajax to allow for editing individual fields.
This was abandoned because the ajax code being used was too buggy or problematic.
Some of the code remained in case the javascript were to be re-added.
This will not happen, so remove the code.