Feature: Provide a custom access denied message on dashboard for accounts created while banner is down (Part 1)
It is entirely possible for this to happen when there is an error contacting the banner database, but for now the most common situations is the 4:00am backup.
Provide a message handling this.

This first part adds the functionality to detect the problem and report it to the user.
The second part will add the ability to manage this problem and other problems for all users on the system.

This requires the following SQL code changes to be applied to a live system:

  start transaction;
  create table mfcs_users (
    user_id bigint not null,
    problems bigint not null default 0,
    date bigint not null,

    primary key (user_id),
    constraint user_id_fkey foreign key (user_id) references users(uid) on delete restrict,

    constraint mfcs_users_user_id_check check (user_id >= 0),
    constraint mfcs_users_problems_check check (problems >= 0)
  );
  grant select, insert, update, delete on mfcs_users to fcs_user;
  commit transaction;

ticket: fcs-9