Bugfix: require request coordinator classification to be defined
A query error happens when request coordinator classification is not defined during a request create operation.
The problem stems from how the form handles the population of the users request classification.
My test account did not have a request classification assigned while I was building a form.
When I attempted to submit, I could not because the field was missing, so I added a classification to the user and attempt to resubmit.
The form showed the classification, but the classification did not make it to the form.