Bugfix: grant/ungrant reviewer role based on venue/room coordinator status in banner database
The code already assigns "Reviewer" role to venue coordinators during the sync if the user does not exist on the system.
However, if the user logs in before the sync occurs, their account does not auto-grant reviewer access.

When the user logs in, during the sync process, process the users reviewer role based on venue/room coordinator status in banner database.
If a problem happens during re-assigning the role, trigger the appropriate problem flag.

ticket: fcs-54