Bugfix: banner user to drupal user conflict due to non-distinctness of banner id in ldap
It appears possible, via ldap's configuration, to have the same banner id for different account names.
This happened when a persons name was changed.
- The old account remained and both have the same banner id.

Help avoid this conflict by requiring support for two new fields in the interface:
1) field_user_is_duplicate: process the isDuplicate ldap field.
2) field_user_primary_account: a boolean setup so that the user is or is not a primary account.

The behavior of two accounts set as primary is undefined at this time.

Force synchronization of accounts and disable non-primary account where possible.
Replace reviewer and proxy information when a primary account is assigned.

Add a watchdog warning in the dashboard for when a user has insufficient access.

Other changes needed to get this done that I do not remember at this time.

ticket: fcs-58