Bugfix: users are not always being granted banner roles, primary issue and others
The users do not always get assigned the banner id, when the database shows that they should be assigned.

The first problem is that if the user_id was not found, a previous user's user id was being incorrectly used.
- The solution is to use a different variable to avoid variable assignment issues.

The second problem is that for old users who existed before the 'primary account' feature was added, under certain conditions they ended up not having their banner roles assigned to them.
- The solution is to forcefully create and assign the primary account setting whenever possible.