Bugfix: allow existing requestor classification to remain even if current editing user is not granted that classification
When I added the ability for a Venue Coordinator to edit an existing application, this ended up causing the request coordinators role to be changed.
This happens because the original design expected the current editor of the form to be the requestor and so the current users status was used to determine which classifications were allowed.
- Managers were also allowed to edit in the original design, but they have access to all classifications.

What happens is that if the current user does not have access to a given classification and that classification is in use on a request, then that classification gets replaced with one of the current editors allowed classifications.
The solution is to preserve the old classification between edits.

ticket: fcs-57