Bugfix: only validate date fields when on a form with the date fields
While testing and resolving ticket fcs-57, I noticed that the 'date is in the past' error was preventing the form from being submitted on pages where the date field did not exist.
Change the behavior to only validate the date field when the date field is on the current forn.