Bugfix: use correct delta on data load so that new revisions are associated with the correct delta
When editing a particular field, such as date, if the delta starts at some value other than 0, then the 0 delta gets re-created and used.
This causes a problem where dates and values do not line up per delta, per revision.

This caused messages on the calendar views to appear, such as:
  "Unable to find 'date' property in for item, skipping. item = stdClass Object ( [id] => 1817 [user_id] => 688 [created] => 1440208400 [updated] => 1442499375 [status] => 3 [step] => 134 [title] => McNeese SAGE [type] => 59 [classification] => 3 [location] => 9 [building] => 18 [room] => 132 [venue_coordinator] => 312 [dates_date] => 1442811600 [mfg_c_0_value] => 1442811600 [dates_date_value] => 1442811600 [dates_time_start] => 1442862000 ) ."

ticket: fcs-62