Bugfix: Banner User Roles not being assigned to some users
This bug was introduced when the user synchronization code was re-design to address ticket: fcs-58.
Manually created user accounts were not having their roles populated and it is very likey that they would not have until the user logged on.

This moves the role and banner role processing code to its own functions so that the updates can be called in hook_user_presave().
This, in part, is done to prevent intinite recursion.

The synchronization code should not have to be called recursively in such a way that user_save() is called multiple times.
This behavior has been corrected and the parameter $allow_recurse has been renamed to $check_for_multiple_accounts.

ticket: fcs-70