Bugfix: invalid e-mails sent because of NULL and 0 comparisions in status check (and related checks)
I was using $status == MFCS_REQUEST_STATUS_DELETED.
- The value for MFCS_REQUEST_STATUS_DELETED is 0.
- If status was NULL when Using == would result in incorrect behavior.

Change == to === to ensure that NULL is not treated the same as 0.

Other checks, such as step, decision, and type, are also affected and have also been updated.

This change requires all of the variables that may be strings representing numbers to be forced to be integers.