Bugfix: strings being treated as integers do not match with === and !===
The drupal database functions, per perhaps PHP, retrieves integers as strings and stores them as strings.
To properly handle these integers, they must be manually cast to an integer.

Some earlier bugfixes dealt with an issue where NULL, 0, and '0' were all treated as the same, causing problems with the e-mails.
The commits are:
- 924c07225b4a7d57c0245c599304f4d0ac9abae5
- e1fafba60ee93f3cca188a825458e24282a835df

Those changes missed some cases that needed to be fixed.

This fixes the cases that I noticed.
I am certain that there are a large number of places where strings are used instead of integers as a result of either drupal or php behavior.
The only time this will become a big problem is when === and !== is used or when 0, NULL, and '0' mean different things.