Update: disable the auto-lock feature for disabled accounts
Having the account still login and then provide a more user-friendly page about the login problems seems to be more effective than trying to use drupals login/validate api and process.
I wasted enough time fighting with it and have to conclude that drupals api is not good for this kind of behavior.

I should follow up this commit with a cron job that will periodically auto-lock accounts that have not been accessed in over 1 year and have no banner roles.