Update: assign effective user id of 1 (system user) for cron jobs
The e-mail system will send e-mails and log the e-mail using the current user id.
When performing cron jobs, the actual user id is 0 (anonymous).

To be more precise, the system should report itself as user id of 1 (system user) when performing cron jobs given that it is not actually anonymous perfoimg the tasks.
This is considered an effective system id and is currently only used and intended for cron jobs.

Only anonymous should be allowed to have an effective system id of 1 (system user) based on the current design.