Update: implement accepted (cancelled), performance improvements, 'all' options
Implement and make sure that requests that have been accepted, but later cancelled are in a different state than those that were cancelled but never accepted.
- This allows for reporting "cancelled" ical requests for ical (important for the ical synchronization process).
- This required extensive changes to be made to the entire project.

Using implode() and then returning the resulting ical string turns out to be incredibly inefficient memory and time wise.
- Return an array and have the caller manually output each array value.
- ~400 calendar entries took about 3 minutes before this change and < 1 minute after this change.

Add support for loading 'all' ical events as opposed to only the current day forward.