Bugfix: requirements forms stopped validating properly when waived
It looks like I was in the middle of some change and got off course and never came back to finish the code.
The code needs to check the new form settings first (which is only valid during the venue available step) and then if none are define check the current waived settings of the form.