Update: only restrict stale submits when the states have been changed
When a request has been changed since the last time the page was loaded, do not submit only when the step has changed.
This prevents the problems and the error messages while not being too restrictive.
Without this change, even comments would cause an error and that could be annoying.

ticket: fcs-74