Feature: allow disabling cancelling support for reviewers and requestors as a whole using a configure setting
This should make it easy to turn on/off cancelling support on request.
The default behavior at this time will be to only allow venue coordinators and managers to cancel.
Requesters will be disbled by default using this feature.

To disable cancelling support, create an array like the following in the settings.php file:
  $conf['disable_cancelling_for'] = array(
    'reviewer' => TRUE,
    'requester' => TRUE,
  );

The following is intended to be the default at this time:
  $conf['disable_cancelling_for'] = array(
    'reviewer' => FALSE,
    'requester' => TRUE,
  );