Bugfix: count options 'returned' should be 'results'
This caused a performance regression in loading the data from the database twice.

The first load was just supposed to be a count() and not include the results.
I am guessing that I renamed the parameter during the development process and overlooked renaming its use in the ajax.inc file.