Bugfix: make sure count queries process distinctness when using distinct_id
The count query should be distinct if the normal query is distinct.
The current design only does the count on the id.

Assigning distinct to the id could cause the query to handle distinctness improperly.
The distinct_id parameter focuses just on the id and therefore makes it safe to add distinct to the count query.

Using the count query on all fields can cause validity problems with the SQL code.
Therefore to get a distinct count of the entire select would require some significant redesign or special handling of drupal cores SQL process.