Update: improve user-specific datatables searching
The searching did not always work as desired.
This may/maynot improve performance by doing the search for user data via a separate query.
The results of that query are then used as user_id based searching.

Some of the functions were not converted to the newer search function and have a different variable structure format.
Some of the functions are exceptional cases and need to search disabled/deleted accounts.